fix minZoom, maxZooms are double

+++ flutter_map_vector/lib/src/theme/layer/fill.dart
@@ -23,13 +23,13 @@ class ThemeFillLayer with _$ThemeFillLayer implements ThemeLayer {
       'number between 0 and 24 inclusive.',
     )
     @JsonKey(name: 'maxzoom')
-    int? maxZoom,
+    double? maxZoom,
     @Assert(
       'minZoom >= 0 && minZoom <= 24',
       'number between 0 and 24 inclusive.',
     )
     @JsonKey(name: 'minzoom')
-    int? minZoom,
+    double? minZoom,
     dynamic paint,
     String? slot,
     String? source,
